requirements.txt
discord.py>=2.0.0
python-binance>=1.0.0
pandas>=1.3.0
ta>=0.10.0
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiohttp-socks>=0.8.0
aiosqlite>=0.19.0
matplotlib>=3.7.0
